Moved code files to src/
