+++ frontend/src/Lighting.jsx
