+++ frontend/src/detailedData.jsx
