+++ frontend/src/diagnosticsTab.jsx
